requirements.txt
django
psycopg2
djangorestframework
coreapi